openpilot v0.9.9 release
date: 2025-06-05T19:54:08
master commit: 8aadf02b2fd91f4e1285e18c2c7feb32d93b66f5

+++ tinygrad_repo/extra/models/transformer.py
@@ -51,7 +51,7 @@ class Transformer:
     maxlen_eye = Tensor.eye(x.shape[1])
     maxlen_eye = maxlen_eye.unsqueeze(0).expand([bs, *maxlen_eye.shape])
 
-    onehot_feat = x.one_hot(self.syms)
+    onehot_feat = x.int().one_hot(self.syms)
 
     onehot = maxlen_eye.cat(onehot_feat, dim=2).flatten(end_dim=1)
